Gone back to using Double for the order quantity when getting a ship estimate, it's pulled from the database in a lot of places so it can't always be a BigDecimal for now.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@590578 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/order/src/org/ofbiz/order/shoppingcart/ShoppingCart.java
@@ -2762,7 +2762,7 @@ public class ShoppingCart implements Serializable {
             if (csii != null && csii.quantity > 0) {
                 if (item.shippingApplies()) {
                     Map itemInfo = item.getItemProductInfo();
-                    itemInfo.put("quantity", BigDecimal.valueOf(csii.quantity));
+                    itemInfo.put("quantity", Double.valueOf(csii.quantity));
                     itemInfos.add(itemInfo);
                 }
             }
